+++ src/exactly_lib/test_case_utils/string_transformer/parse_string_transformer.py
@@ -5,7 +5,7 @@ from exactly_lib.test_case_utils.expression.parser import GrammarParsers
 from exactly_lib.test_case_utils.string_transformer import names
 from exactly_lib.test_case_utils.string_transformer import sdvs
 from exactly_lib.test_case_utils.string_transformer.impl import filter, replace, sequence, identity, case_converters, \
-    tcds_paths_replacement, strip_trailing_new_lines, strip_trailing_space
+    tcds_paths_replacement, strip_trailing_new_lines, strip_trailing_space, strip_space
 from exactly_lib.test_case_utils.string_transformer.impl.run_program import parse as parse_run
 from exactly_lib.util.cli_syntax.elements import argument as a
 from exactly_lib.util.name_and_value import NameAndValue
@@ -61,6 +61,11 @@ GRAMMAR = grammar.Grammar(
             names.STRIP_TRAILING_SPACE,
             grammar.Primitive(strip_trailing_space.parse,
                               strip_trailing_space.SyntaxDescription())
+        ),
+        NameAndValue(
+            names.STRIP_SPACE,
+            grammar.Primitive(strip_space.parse,
+                              strip_space.SyntaxDescription())
         ),
         NameAndValue(
             names.RUN_PROGRAM_TRANSFORMER_NAME,