+++ test/exactly_lib_test/test_case_utils/string_transformers/test_resources/argument_syntax.py
@@ -86,5 +86,9 @@ def strip_trailing_space() -> str:
     return names.STRIP_TRAILING_SPACE
 
 
+def strip_space() -> str:
+    return names.STRIP_SPACE
+
+
 def tcds_path_replacement() -> str:
     return names.TCDS_PATH_REPLACEMENT