+++ CHANGELOG.md
@@ -21,6 +21,7 @@ with exception of a "0." prefix until version 1 is released.
  - Type `line-matcher` - `contents` - matches text contents using a `string-matcher`
  - Type `string-transformer` - `strip-trailing-new-lines`
  - Type `string-transformer` - `strip-trailing-space`
+ - Type `string-transformer` - `strip-space`
  - Type `integer-matcher`
 
 ### Changed